--- app/src/main/java/com/asfoundation/wallet/repository/EskillsGamesRepositoryType.kt
@@ -1,9 +1,0 @@
-package com.asfoundation.wallet.repository
-
-import com.appcoins.wallet.ui.widgets.GameData
-import io.reactivex.Single
-
-interface EskillsGamesRepositoryType {
-
-    fun getEskillsGamesListing(): Single<List<GameData>>
-}
