+++ app/src/main/java/com/asfoundation/wallet/repository/GamesRepository.kt
@@ -3,6 +3,7 @@ package com.asfoundation.wallet.repository
 import android.util.Log
 import com.appcoins.wallet.core.network.backend.api.GamesApi
 import com.appcoins.wallet.core.network.eskills.api.AppDataApi
+import com.appcoins.wallet.core.network.eskills.api.EskillsGamesApi
 import com.appcoins.wallet.ui.widgets.GameData
 import com.appcoins.wallet.ui.widgets.GameDetailsData
 import com.appcoins.wallet.ui.widgets.Screenshot
@@ -12,14 +13,20 @@ import javax.inject.Inject
 
 @BoundTo(supertype = GamesRepositoryType::class)
 class GamesRepository @Inject constructor(
-  private val gamesApi: GamesApi,
+  private val gamesApi: EskillsGamesApi,
   private val appDataApi: AppDataApi) :
   GamesRepositoryType {
 
   private val defaultBackground = "https://image.winudf.com/v2/image1/Y29tLm5hdGhuZXR3b3JrLnVsdHJhcHJvX3NjcmVlbl8xXzE2MzE3MzE3MTlfMDQ5/screen-1.webp?fakeurl=1&type=.webp"
 
   override fun getGamesListing(): Single<List<GameData>> {
-    return gamesApi.getGamesListing()
+    return gamesApi.getGamesListing(
+      group = "e-skills",
+      limit = "20",
+      sort = "downloads7d",
+      language = "en",
+      store = "apps"
+    )
       .map { it.dataList.list.map {
         Log.d("App Name","Name: "+it.appName)
         Log.d("App Icon","Icon: "+it.appIcon)
